backend/requirements.txt
flask>=3.0,<4
python-dotenv>=1.0,<2
flask-cors>=4.0,<6
openai>=1.40,<2
pinecone>=5.0,<7
twilio>=9.0,<10
pydantic>=2.0,<3
flask-sqlalchemy>=3.1,<4
PyJWT>=2.8,<3
email-validator>=2.0,<3
pypdf>=4.0,<6
fpdf2>=2.7,<3
Pillow>=10.0,<12
requests>=2.31,<3
